Visibility of command constants was changed. The lock command.  - constants which describe commands that supposed to be common for all the platforms  have package-visibility now.   - The argument preparation for the lockDevice command was moved to the MobileCommand.
