+++ src/main/java/io/appium/java_client/android/AndroidMobileCommandHelper.java
@@ -281,12 +281,9 @@ public class AndroidMobileCommandHelper extends MobileCommand {
     }
 
     /**
-     * This method forms a {@link java.util.Map} of parameters for the
-     * device locking.
-     *
-     * @return a key-value pair. The key is the command name. The value is a
-     * {@link java.util.Map} command arguments.
+     * This method was moved to {@link MobileCommand#hideKeyboardCommand(String, String)}
      */
+    @Deprecated
     public static Map.Entry<String, Map<String, ?>>  lockDeviceCommand() {
         return new AbstractMap.SimpleEntry<>(LOCK, prepareArguments("seconds", 0));
     }
