+++ src/main/java/io/appium/java_client/android/LocksAndroidDevice.java
@@ -17,8 +17,8 @@
 package io.appium.java_client.android;
 
 import static io.appium.java_client.android.AndroidMobileCommandHelper.isLockedCommand;
-import static io.appium.java_client.android.AndroidMobileCommandHelper.lockDeviceCommand;
 import static io.appium.java_client.android.AndroidMobileCommandHelper.unlockCommand;
+import static io.appium.java_client.MobileCommand.lockDeviceCommand;
 
 import io.appium.java_client.CommandExecutionHelper;
 import io.appium.java_client.ExecutesMethod;
@@ -37,7 +37,7 @@ public interface LocksAndroidDevice extends ExecutesMethod {
      * This method locks a device.
      */
     default void lockDevice() {
-        CommandExecutionHelper.execute(this, lockDeviceCommand());
+        CommandExecutionHelper.execute(this, lockDeviceCommand(0));
     }
 
     /**
